+++ NEWS.md
@@ -3,7 +3,10 @@
 * `daemons()` interface has been simplified with a single `value` argument along with `...` for passing additional parameters.
   + Allows supplying the argument `nodes` for running an active queue (task scheduler).
   + Calling with no arguments now provides an improved view of the current number of connections / daemons / nodes, replacing the previous `daemons("view")` functionality.
-* `server()` gains the argument `nodes` for launching an active queue with the specified number of nodes, and `max_tasks` for limiting the number of tasks to be executed before exiting.
+* `server()` gains the following arguments:
+  + `nodes` for launching an active queue with the specified number of nodes.
+  + `tasks` for specifying the maximum number of tasks before exiting.
+  + `runtime` for specifying a soft walltime before exiting.
 * `eval_mirai()` is now removed as an alias for `mirai()`.
 * Invalid URLs provided to `server()` and `daemons()` now error and return in a timely fashion instead of potentially causing a hang.
 * Requires nanonext >= 0.7.3.9002 and R >= 3.5.